requirements.txt
# Hot Tub Controller dependencies
python-kasa==0.5.2
requests>=2.31.0
aiohttp>=3.9.0